--- front/src/components/Carrito.js
@@ -1,120 +1,0 @@
-import React, { Fragment, useState} from 'react'
-import MetaData from './layout/MetaData'
-import { Link } from 'react-router-dom'
-import {numberWithCommas} from '../numberWithCommas'
-
-export const Carrito = () => {
-    const [quantity, setQuantity] = useState(1)
-
-    const increaseQty = () => {
-        const contador = document.querySelector('.count')
-        const qty = contador.valueAsNumber+1;
-        setQuantity(qty)
-     }
-  
-     const decreaseQty = () => {
-      const contador = document.querySelector('.count')
-  
-      const qty = contador.valueAsNumber-1;
-      setQuantity(qty)
-   }
-
-    //Json de ejemplo
-   let cartItems=[
-        {
-            "_id": "6351a0ac9dc6821674880d59",
-            "nombre": "Plástico PET Cristal",
-            "precio": 2815000,
-            "imagen": "./images/PETCristal.jpg",
-            "inventario": 20,
-        },
-        {
-            "_id": "63519e876757fd7f1330ca64",
-            "nombre": "Aluminio por tonelad",
-            "precio": 5111000,
-            "imagen": "./images/Aluminio.jpg",
-            "inventario": 5,
-        },
-        {
-            "_id": "63519fc16757fd7f1330ca67",
-            "nombre": "Plástico Rígido por Tonelada",
-            "precio": 1911000,
-            "imagen": "./images/PlasticoRigido.jpg",
-            "inventario": 3,
-        },
-    
-    ]
-
-cartItems = Array.from(cartItems);
-
-    return (
-        <Fragment>
-            <MetaData title={'Your Cart'} />
-            
-
-            {cartItems.length === 0 ? <h2 className="mt-5">Su carrito esta vacio</h2> : (
-                <Fragment>
-                    
-                    <h2 className="mt-5">Su Carrito: <b>{cartItems.length} items</b></h2>
-
-                    <div className="row d-flex justify-content-between">
-                        <div className="col-12 col-lg-8">
-
-                        {cartItems && cartItems.map (item => (
-                                <Fragment>
-                                    <hr />
-
-                                    <div className="cart-item" key={item.nombre}>
-                                        <div className="row">
-                                            <div className="col-4 col-lg-3">
-                                                <img src={item.imagen} alt={item.nombre} height="90" width="115" />
-                                            </div>
-                                            <div className="col-5 col-lg-3">
-                                                <Link to={`/producto/${item._id}`}>{item.nombre}</Link>
-                                            </div>
-                                            <div className="col-4 col-lg-2 mt-4 mt-lg-0">
-                                                <p id="card_item_price">${numberWithCommas(item.precio)}</p>
-                                            </div>
-
-                                            <div className="col-4 col-lg-3 mt-4 mt-lg-0">
-                                                <div className="stockCounter d-inline">
-                                                    <span className="btn btn-danger minus" onClick={decreaseQty}>-</span>
-
-                                                    <input type="number" className="form-control count d-inline" value={quantity} readOnly />
-
-                                                    <span className="btn btn-primary plus" onClick={increaseQty}>+</span>
-                                                </div>
-                                            </div>
-
-                                            <div className="col-4 col-lg-1 mt-4 mt-lg-0">
-                                                <i id="delete_cart_item" className="fa fa-trash btn btn-danger" ></i>
-                                            </div>
-
-                                        </div>
-                                    </div>
-                                    <hr />
-                                </Fragment>
-                            
-                        ))}
-                        </div>
-
-                        <div className="col-12 col-lg-3 my-4">
-                            <div id="order_summary">
-                                <h4>Total de la Compra</h4>
-                                <hr />
-                                <p>Subtotal:  <span className="order-summary-values">$9.837.000</span></p>
-                                <p>Transporte: <span className="order-summary-values">$50.000</span></p>
-                                <p>Total: <span className="order-summary-values">$9.887.000</span></p>
-
-                                <hr />
-                                <button id="checkout_btn" className="btn btn-primary btn-block">Comprar!</button>
-                            </div>
-                        </div>
-                    </div>
-                </Fragment>
-            )}
-        </Fragment>
-    )
-}
-
-export default Carrito
