+++ front/src/components/admin/Dashboard.js
@@ -47,7 +47,7 @@ export const Dashboard = () => {
                                 <div className="col-xl-3 col-sm-6 mb-3">
                                     <div className="card text-white bg-danger o-hidden h-100">
                                         <div className="card-body">
-                                            <div className="text-center card-font-size">Ventas<br /> <b>34</b></div>
+                                            <div className="text-center card-font-size">Pedidos<br /> <b>34</b></div>
                                         </div>
                                         <Link className="card-footer text-white clearfix small z-1" to="/admin/orders">
                                             <span className="float-left">Ver Detalles</span>
