+++ front/src/constants/productConstants.js
@@ -2,8 +2,8 @@ export const ALL_PRODUCTS_REQUEST ="ALL_PRODUCTS_REQUEST"
 export const ALL_PRODUCTS_SUCCESS="ALL_PRODUCTS_SUCCESS"
 export const ALL_PRODUCTS_FAIL="ALL_PRODUCTS_FAIL"
 
-export const PRODUCT_DETAILS_REQUEST="ALL_DETAILS_REQUEST"
-export const PRODUCT_DETAILS_SUCCESS="ALL_DETAILS_SUCESS"
-export const PRODUCT_DETAILS_FAIL="ALL_DETAILS_FAIL"
+export const PRODUCT_DETAILS_REQUEST="PRODUCT_DETAILS_REQUEST"
+export const PRODUCT_DETAILS_SUCCESS="PRODUCT_DETAILS_SUCCESS"
+export const PRODUCT_DETAILS_FAIL="PRODUCT_DETAILS_FAIL"
 
 export const CLEAR_ERRORS="CLEAR_ERRORS"
