--- front/src/constants/saleConstants.js
@@ -1,10 +1,0 @@
-export const ALL_SALES_REQUEST ="ALL_SALES_REQUEST"
-export const ALL_SALES_SUCCESS="ALL_SALES_SUCCESS"
-export const ALL_SALES_FAIL="ALL_SALES_FAIL"
-
-export const SALE_DETAILS_REQUEST="ALL_DETAILS_REQUEST"
-export const SALE_DETAILS_SUCCESS="ALL_DETAILS_SUCESS"
-export const SALE_DETAILS_FAIL="ALL_DETAILS_FAIL"
-
-
-export const CLEAR_ERRORS="CLEAR_ERRORS"
