+++ front/src/store.js
@@ -1,15 +1,12 @@
 import { createStore, combineReducers, applyMiddleware} from 'redux';
 import thunk from 'redux-thunk';
 import { composeWithDevTools} from 'redux-devtools-extension';
-import { productsReducer, productDetailsReducer} from './reducer/productReducer';
-import {salesReducer} from './reducer/saleReducer';
+import { productsReducer, productDetailsReducer } from './reducer/productReducer';
 
 const reducer= combineReducers ({
     products:productsReducer,
-    productDetails: productDetailsReducer,
-    sales:salesReducer,
-   })
-
+    productDetails: productDetailsReducer
+})
 
 let initialState = {}
